Fix schema mismatch
Updated MySQL-Catalog-schema.sql to match the Hibernate class
definitions.  Restored the scripts that add default values for
CREATION_TIMESTAMP columns.  Reverted an attempted fix to a
foreign key definition for the VNF_RES_CUSTOM_TO_VF_MODULE_CUSTOM
table that unintentionally added an extra column to the table.

Issue: SO-180
Change-Id: Ie76059f4c2497060a4fba8744e974b29d33b16f1
Signed-off-by: Rob Daugherty <rd472p@att.com>