docker-config for local environment
We can use docker-compose to bring up SO containers in
a development environment.

This solution is lighter weight than kubernetes with
rancher, and should be a lot easier to use for those who
don't need all the management functions provided by
kubernetes.

I've included a script that fetches the springboot
overlay files (config files) from the oom repo and
parameterizes them for the docker-config environment.
This should alleviate the issues we would otherwise
have with keeping two sets of overlay files in sync.

Note that nothing here affects any delivered artifact,
so perhaps the PTL will consider merging it into
casablanca, even though it is past the maintenance
release date.

Change-Id: I0deeb853484fdfd0862b02121e0922fe7fc923e4
Signed-off-by: Rob Daugherty <rd472p@att.com>
Issue-ID: SO-1374